feat: bookCart Quantity Integer 형 변경, DTO 정리

+++ src/main/java/com/nhnacademy/bookstoreback/bookcart/domain/dto/request/UpdateBookCartRequest.java
@@ -1,4 +1,4 @@
 package com.nhnacademy.bookstoreback.bookcart.domain.dto.request;
 
-public record UpdateBookCartRequest(Long bookQuantity) {
+public record UpdateBookCartRequest(Integer bookQuantity) {
 }
